moved onboarding stuff into views package
